+++ packages/app/index.ts
@@ -271,7 +271,7 @@ async function init() {
   if (!app.commandLine.hasSwitch("disable-bring-to-top-on-focus")) {
     main.window.on("focus", () => {
       if (main.location === "/") {
-        accounts.getSelectedAccount().instance.gmail.view.webContents.focus();
+        accounts.getSelectedAccount().instance.gmail.viewOrNull?.webContents.focus();
       }
     });
   }